Changed the variable type of count_button_press and count_button_rel from uint8_t to uint16_t to manage the LONG PRESSION CONDITION of the PushButton component, but there is still the problem of changing the parameter to be modified with the potentiometer (SINGLE CLICK CONDITION)

+++ AY2021_Project_2_Workspace/Project_2.cydsn/Global.h
@@ -62,8 +62,8 @@
     extern uint8_t seconds;
     extern uint8_t minutes;
     extern uint8_t hours;
-    extern uint8_t count_button_press;
-    extern uint8_t count_button_rel;
+    extern uint16_t count_button_press;
+    extern uint16_t count_button_rel;
     extern uint8_t count_blinking;
     
     /* Variable related to the selection of the different parameters that is possible to change in the configuration mode
